Moving files to a single one
Easier to manipulate and read
