--- module_anime/models/__init__.py
@@ -1,2 +1,0 @@
-from .Anime import Anime
-from .AnimeTitles import AnimeTitles
